requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate
Flask-Login
Flask-APScheduler
PyMySQL==1.1.0
paramiko==3.4.0
python-gitlab==4.4.0
python-dotenv==1.0.1
cryptography==42.0.5
openpyxl
